API

Plotting recipes

ItemResponsePlots.item_characteristic_curve — Function

item_characteristic_curve(model::ItemResponseModel, item, response)
item_characteristic_curve(model::ItemResponseModel, item)

Create a plot of the item characteristic curve for response of item.

An item characteristic curve plots the probability for response of item given the person ability θ.

If response is omitted, the default plot behaviour depends on model:

For models where response_type(model) == Dichotomous the item characteristic curve for response = 1 is plotted, i.e. the probability of a correct response.

For models where response_type(model) <: Union{Nominal, Ordinal} all category characteristic curves are plotted.

Plot attributes

Generic

color: The color of the item characteristic curve.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the item characteristic curve. default: -3.0:0.01:3.0.

show_data: Overlay observed data in the plot. Ability estimates are aggregated according to bins. For models with estimation_type(model) == SamplingEstimate the posterior means of ability estimates are used for binning. default: false.

bins: The number of bins for the observed data. default: 6.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.item_characteristic_curve! — Function

item_characteristic_curve(model::ItemResponseModel, item, response)
item_characteristic_curve(model::ItemResponseModel, item)

Create a plot of the item characteristic curve for response of item.

An item characteristic curve plots the probability for response of item given the person ability θ.

If response is omitted, the default plot behaviour depends on model:

For models where response_type(model) == Dichotomous the item characteristic curve for response = 1 is plotted, i.e. the probability of a correct response.

For models where response_type(model) <: Union{Nominal, Ordinal} all category characteristic curves are plotted.

Plot attributes

Generic

color: The color of the item characteristic curve.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the item characteristic curve. default: -3.0:0.01:3.0.

show_data: Overlay observed data in the plot. Ability estimates are aggregated according to bins. For models with estimation_type(model) == SamplingEstimate the posterior means of ability estimates are used for binning. default: false.

bins: The number of bins for the observed data. default: 6.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.item_information_curve — Function

item_information_curve(model::ItemResponseModel, item, response)
item_information_curve(model::ItemResponseModel, item)

Create a plot of the item information curve for response of item.

If response is omitted, the default plot behaviour depends on model:

For models where response_type(model) == Dichotomous the item information curve for response = 1 is plotted.

For models where response_type(model) <: Union{Nominal, Ordinal} all category information curves are plotted.

Plot attributes

Generic

color: The color of the item information curve.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the item information curve. default: -3.0:0.01:3.0.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.item_information_curve! — Function

item_information_curve(model::ItemResponseModel, item, response)
item_information_curve(model::ItemResponseModel, item)

Create a plot of the item information curve for response of item.

If response is omitted, the default plot behaviour depends on model:

For models where response_type(model) == Dichotomous the item information curve for response = 1 is plotted.

For models where response_type(model) <: Union{Nominal, Ordinal} all category information curves are plotted.

Plot attributes

Generic

color: The color of the item information curve.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the item information curve. default: -3.0:0.01:3.0.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.expected_score_plot — Function

expected_score_plot(model::ItemResponseModel)
expected_score_plot(model::ItemResponseModel, items)

Create a plot of the expected score for model.

If items is specified, the expected score is plotted according to the subtest including only items. If items is omitted, the expected score is plotted for all items included in model.

Plot attributes

Generic

color: The color of the expected score plot.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the expected scores. default: -3.0:0.01:3.0.

scoring_function: The scoring function applied to the expected scores.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000.

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.expected_score_plot! — Function

expected_score_plot(model::ItemResponseModel)
expected_score_plot(model::ItemResponseModel, items)

Create a plot of the expected score for model.

If items is specified, the expected score is plotted according to the subtest including only items. If items is omitted, the expected score is plotted for all items included in model.

Plot attributes

Generic

color: The color of the expected score plot.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the expected scores. default: -3.0:0.01:3.0.

scoring_function: The scoring function applied to the expected scores.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000.

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.information_plot — Function

information_plot(model::ItemResponseModel)
information_plot(model::ItemResponseModel, items)

Create a plot of the test information for model

If items is specified, the test information is plotted according to the subtest including only items. if items is omitted, the test information is plotted for all items included in model.

Plot attributes

Generic

color: The color of the information plot.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the information. default: -3.0:0.01:3.0.

scoring_function: The scoring function applied to the information.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000.

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

ItemResponsePlots.information_plot! — Function

information_plot(model::ItemResponseModel)
information_plot(model::ItemResponseModel, items)

Create a plot of the test information for model

If items is specified, the test information is plotted according to the subtest including only items. if items is omitted, the test information is plotted for all items included in model.

Plot attributes

Generic

color: The color of the information plot.

uncertainty_color: The color of the displayed uncertainty information. For plots with uncertainty intervals this is the color of the confidence band. For plots with sample based uncertainty information this is the line color of the samples.

theta: The values of theta for which to plot the information. default: -3.0:0.01:3.0.

scoring_function: The scoring function applied to the information.

Specific

Models with SamplingEstimate

samples: The number of samples to plot. default: 1000.

uncertainty_type: Changes how the uncertainty of the estimate is displayed. If uncertainty_type = :samples, then iterations from the MCMC estimation are plotted. If unvertainty_type = :interval, then uncertainty intervals are plotted. default: :samples

quantiles: The lower and upper quantile for uncertainty intervals. default: (0.1, 0.9)

aggregate_fun: A function that aggregates MCMC samples. The provided function must take a vector as input and output a scalar value. If aggregate_fun = nothing no aggregate is plotted. default: mean

Plots

ItemResponsePlots.itemplot — Function

itemplot(model::ItemResponseModel, item, args...; kwargs...)

Create an item plot for item of model.

The resulting item plot contains the item characteristic curve (left) and the item information curve (right).

The additional args... and kwargs... are passed to the lower level functions item_characteristic_curve and item_information_curve.

ItemResponsePlots.testplot — Function

testplot(model::ItemResponseModel, args...; kwargs...)

Create a plot for test characteristics of model.

The resulting plot contains the test characteristic curve/expected scores (left) and the test information curve (right).

The additional args... and kwargs... are passed to the lower level plotting functions expected_score_plot and information_plot.

Settings

ItemResponsePlots.getdefault — Function

getdefault(attribute::String)

Fetch the default value for attribute.

ItemResponsePlots.setdefault! — Function

setdefault!(attribute::String, value)

Set a new default value for attribute.